test: consolidate duplicated fixture/expectAlert helpers into test/support
The fixture() path builder and expectAlert() dialog helper were copy-pasted
across all four Playwright spec files, risking silent drift if only some
copies were updated.
