+++ test/specs/aync-create.js
@@ -1,8 +1,5 @@
-const path = require('path');
 const { test, expect } = require('@playwright/test');
-
-const JQUERY_VERSION = process.env.JQUERY_VERSION || '4.0.0';
-const fixture = (name) => 'file://' + path.join(process.cwd(), 'test/integration/html', 'jquery-' + JQUERY_VERSION, name);
+const { fixture } = require('../support/helpers');
 
 test.describe('Test async create', () => {
   test('should render async created context menu', async ({ page }) => {
