+++ test/specs/callback.js
@@ -1,24 +1,11 @@
-const path = require('path');
 const { test, expect } = require('@playwright/test');
-
-const JQUERY_VERSION = process.env.JQUERY_VERSION || '4.0.0';
-const fixture = (name) => 'file://' + path.join(process.cwd(), 'test/integration/html', 'jquery-' + JQUERY_VERSION, name);
+const { fixture, expectAlert } = require('../support/helpers');
 
 async function openCallbackMenu(page) {
   await page.goto(fixture('callback.html'));
   await page.click('.context-menu-one', { button: 'right' });
   await page.waitForSelector('#context-menu-layer');
   await expect(page.locator('.context-menu-root')).toBeVisible();
-}
-
-async function expectAlert(page, action, message) {
-  let dialogMessage = null;
-  page.once('dialog', async (dialog) => {
-    dialogMessage = dialog.message();
-    await dialog.accept();
-  });
-  await action();
-  expect(dialogMessage).toBe(message);
 }
 
 test.describe('Test callback', () => {
